colors now as U8 type

+++ rialto_viewer/web/cesium/cloud_shape.dart
@@ -7,9 +7,9 @@ part of rialto.viewer;
 class CloudShape extends Shape {
     int numPoints;
     Float32List points;
-    Float32List colors;
+    Uint8List colors;
 
-    CloudShape(Float32List this.points, Float32List this.colors) : super("cloud") {
+    CloudShape(Float32List this.points, Uint8List this.colors) : super("cloud") {
         numPoints = points.length ~/ 3;
         assert(numPoints * 3 == points.length);
         assert(numPoints * 4 == colors.length);
